Implement q_new and q_free functions
The q_new function is responsible for adding a queue with both prev and next pointers
pointing to the queue's beginning.
The q_free function is responsible for releasing all storage space on the queue.

For q_new, first allocate space of the size of struct list_head.
If allocation fails, return NULL. If successful, use the INIT_LIST_HEAD function
to set prev and next pointers to itself, then return the head node.

For q_free, first check if the head exists; if not, do nothing.
Here, an indirect pointer to the beginning of the queue's next pointer is used.
The target pointer is used to release node space. The advantage of using an indirect pointer
is that it doesn't need to move between nodes; changing the value of
the indirect pointer can make the next pointer on the node point to the next node.
Finally, release the space of the head node.