Implement q_swap and q_reverse
Implementing q_swap function to swap nodes in the queue under the conditions that
the queue is not empty and there is more than one node. Use prev and curr pointers to point to
head->next and prev->next respectively. Record the pointers of prev->prev and curr->next,
then swap the positions of nodes pointed by prev and curr.

Implementing q_reverse function to reverse the queue under the conditions that the queue
is not empty and there is more than one node. Pairwise swap the head node with the tail node
and reverse their positions.